#6: Initial attempt at project restructuring.
