+++ Assets/Capsule.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 5425691ad8b01419ebaca597d91ca887
+folderAsset: yes
+timeCreated: 1510594549
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
