+++ Assets/Capsule/HE.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 2b40b83e34136494a8685b7a97dfe57d
+folderAsset: yes
+timeCreated: 1510594549
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
