+++ Assets/Capsule/Mpc.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: bdfdfd5ee99e949db927dee3fb49d7ac
+folderAsset: yes
+timeCreated: 1510594549
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
