+++ Assets/Capsule/Mpc/Rss.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 71fc5e77a8c14421f94c4c086e91c359
+folderAsset: yes
+timeCreated: 1510594549
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
