+++ Assets/Capsule/Mpc/Spdz.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 21b53c0b77fb04ca59ce9bc645934865
+folderAsset: yes
+timeCreated: 1510594549
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
