+++ Assets/Common.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: ac206dce553334a56af76fca99c5408d
+folderAsset: yes
+timeCreated: 1510594545
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
