+++ Assets/Network.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 20b03b0926cfa4941b45816bdd169fc8
+folderAsset: yes
+timeCreated: 1510594545
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
