+++ Assets/Network/Command.cs
@@ -1,0 +1,19 @@
+
+namespace OpenMined.Network
+{
+
+    public class Command
+    {
+        // given that SyftController keeps lists of objects of base types
+        // (at the time of writing this is only Tensors) then this command
+        // selects one of these generic types and performs a command.
+        public string objectType;
+        // i.e. "tensor"
+        public int objectIndex;
+        //
+
+        // name of the function to be called
+        public string functionCall;
+
+    }
+}
