+++ Assets/Network/NetMqPublisher.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: a9fbbe2ba3e6940bda392749d3e0a578
+timeCreated: 1510595644
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
