+++ Assets/Network/SyftServer.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 1aaf87c4f05f94adaaa3ec9981fd9470
+timeCreated: 1510594773
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
