--- Assets/ServerObject.cs
@@ -1,100 +1,0 @@
-﻿using System.Diagnostics;
-using System.Threading;
-using NetMQ;
-using NetMQ.Sockets;
-using UnityEngine;
-
-public class NetMqPublisher
-{
-    private readonly Thread _listenerWorker;
-
-    private bool _listenerCancelled;
-
-    public delegate string MessageDelegate(string message);
-
-    private readonly MessageDelegate _messageDelegate;
-
-    private readonly Stopwatch _contactWatch;
-
-    private const long ContactThreshold = 1000;
-
-    public bool Connected;
-
-    private void ListenerWork()
-    {
-        AsyncIO.ForceDotNet.Force();
-        using (var server = new ResponseSocket())
-        {
-            server.Bind("tcp://*:12346");
-
-            while (!_listenerCancelled)
-            {
-                Connected = _contactWatch.ElapsedMilliseconds < ContactThreshold;
-                string message;
-                if (!server.TryReceiveFrameString(out message)) continue;
-                _contactWatch.Restart();
-                var response = _messageDelegate(message);
-                server.SendFrame(response);
-            }
-        }
-        NetMQConfig.Cleanup();
-    }
-
-    public NetMqPublisher(MessageDelegate messageDelegate)
-    {
-        _messageDelegate = messageDelegate;
-        _contactWatch = new Stopwatch();
-        _contactWatch.Start();
-        _listenerWorker = new Thread(ListenerWork);
-    }
-
-    public void Start()
-    {
-        _listenerCancelled = false;
-        _listenerWorker.Start();
-    }
-
-    public void Stop()
-    {
-        _listenerCancelled = true;
-        _listenerWorker.Join();
-    }
-}
-
-public class ServerObject : MonoBehaviour
-{
-    public bool Connected;
-    private NetMqPublisher _netMqPublisher;
-    private string _response;
-
-	private SyftController controller;
-
-	[SerializeField]
-	private ComputeShader shader;
-
-    private void Start()
-    {
-        _netMqPublisher = new NetMqPublisher(HandleMessage);
-        _netMqPublisher.Start();
-
-		controller = new SyftController (shader);
-    }
-
-    private void Update()
-    {
-        var position = transform.position;
-//		_response = position.x
-        Connected = _netMqPublisher.Connected;
-    }
-
-    private string HandleMessage(string message)
-    {
-        // Not on main thread
-		return controller.processMessage (message);;
-    }
-
-    private void OnDestroy()
-    {
-        _netMqPublisher.Stop();
-    }
-}
