+++ Assets/Syft.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 24a97a9ce581841a985a8d022c1dd643
+folderAsset: yes
+timeCreated: 1510594546
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
