+++ Assets/Syft/Math.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 6a63fc47448d34dcba712694a14e9e57
+folderAsset: yes
+timeCreated: 1510594549
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
