+++ Assets/Syft/Math/Shaders.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 609ff5eca7fd74990a18693dd1cf8cbc
+folderAsset: yes
+timeCreated: 1510596431
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
