+++ Assets/Syft/Tensor.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: a4c512d75767c4ee087ca6f9cfdb5972
+folderAsset: yes
+timeCreated: 1510594546
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
