+++ Assets/Syft/Tensor/FloatTensor.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: b53afe9558e2d43aeab56828b5cd62e2
+timeCreated: 1510595997
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
